+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - import code
 - add tests
+- add --root flag
+- add proper integration tests
 
 ## [0.0.1] (unreleased)